Separate PositionItem
